Finised move to proxy/optimizers
