+++ python/ray/rllib/agents/sac/proxy/optimizers/sync_replay_optimizer.py
@@ -3,11 +3,13 @@ import collections
 import numpy as np
 
 import ray
-from ray.rllib.agents.sac.sac.rllib_proxy._patched._replay_buffer import (
-    ReplayBuffer, PrioritizedReplayBuffer)
+from ray.rllib.agents.sac.proxy.optimizers.replay_buffer import (
+    ReplayBuffer, PrioritizedReplayBuffer
+)
 from ray.rllib.evaluation.metrics import get_learner_stats
-from ray.rllib.evaluation.sample_batch import SampleBatch, DEFAULT_POLICY_ID, \
-    MultiAgentBatch
+from ray.rllib.evaluation.sample_batch import (
+    SampleBatch, DEFAULT_POLICY_ID, MultiAgentBatch
+)
 from ray.rllib.optimizers.policy_optimizer import PolicyOptimizer
 from ray.rllib.utils.annotations import override
 from ray.rllib.utils.compression import pack_if_needed
